fix adc real-time port bug
The sync out port on the real-time status ports would not draw if already
configured without MTS enabled. This adds a state check on to see if MTS has
changed to determine if the mask needs to be updated.

Could consider changing the inclusoin of the real-time status ports as a mask
callback function instead.