zcu216 support for iWave quad qsfp fmc+ module
The quad qsfp fmc+ module extends the number of qsfp ports the zcu216 can have
and enables use of two simultaneous 100GbE CMACs. Only two of the four ports are
conencted to quad GTs on the zcu216. Port 0 remains the 4x25 sfp cage. Port 1
and port 2 two now correspond to the two usable ports on the quad fmc+ module.
See the platform file for more information for which ports are used and how
physical ports are mapped to cmac locations.