Remove `bus_expander` on `rx_valid` under the mask
A `bus_expander` and logical `or` on the `rx_valid` output prevented compiling
100G when the `rx_valid` was driving the user design. Removing those blocks the
100G builds and operates as expected.

Not sure what the origin, or why, the bus expand was present. Most likely is a
carry over from the 40g block implementation since the 100g mask and callbacks
inherited from 40g. There is even still callbacks in the mask parameters that
use the 40g version instead of the 100g version...