Initial commit for port to vitis model composer
This commit is initial work to start porting from System Generator to
Vitis Model Composer (VMC). Current efforts _only_ support RFSoC platforms.
The toolflow has been ran with a compilation that builds successfully
with hardware equivalent to what is present in the typical CASPER
tutorial 1. Future work will need to extend support to other relevant
CASPER platforms.

VMC introduces the ability to analyze, implement, and export
hardware at the "subsystem" level, independent of other subsystems.
This initial work _does not_ take advantage of this capability. Rather,
the "Legacy HDL Flow" is adapted to maintain a single top-level
subsystem hierarhcy that can be exported to HDL. This functions
equivalently to what was familiar with System Generator. As a
consequence, the new "hardware flow" and other capabilities of model
composer are not available. This does not impact any previous CASPER
integrated capabilities. All capabilities are expected to remain intact
and function equivalently to what previously done with System Generator.

There continue to be changes made by updates made to VMC that change the
way the HDL underlying blockset parameters are configured. Specifically,
the `gui_display_data_type` must now be set before an `airth_type` of
`Boolean` can be selected. This was change for the snapshot block's
`assert` block. This change is explained as a warning to recall that
similar issues were needed when first moving to versions of Vivado that
warned of deprecation to System Generator. More changes should be
expected to need to be discovered, and fixed.

In this port to Model Composer, individual RFSoC platforms are removed
in favor of a single `RFSoC` yellow block where the individual RFSoC
platform is selected. This is experimental, but should simplify
maintaining and porting RFSoC platforms. The drawback is that this may
limit platform specific configuration. However, it is expected that most
of this can be reconciled in the individual platform's yellow block
python generator file.