port red pitaya and snap to model composer
Red Pitaya and SNAP now use the vmc hub and single configuration mask
script. The Red Pitaya platform block's mask also shows an example of
using the vmc configuration mask for initialization and another callback
for additional configuration and setup.

The SNAP compiles _only_ without the microblaze. Meaning, control must
be done using a raspberry pi. Additional work needs to be done to
re-enable use of microbalze for communication control. The
`use_microblaze` option remains present in the mask as a hidden
parameter for when the work is done to use microblaze.