Purge platform blocks
The approach will be to go back and add platforms back when there is a
need to target them. If a platform is not currently supported in the
tools it should not be an available option. For example, platforms like
VCU118/128 require the microblaze core for control communications.
However, the microblaze hasn't been reliably updated since Vivado
2021.1/2. The idea therefore is to treat brining platforms back as a
porting process to fix and update required infrastructure and support.

SNAP is still present because a Raspberry Pi can be used for control
instead of the microblaze.